Fix sign comparison warnings

+++ core/src/OpenMP/Kokkos_OpenMP_Parallel_For.hpp
@@ -401,7 +401,7 @@ class ParallelFor<FunctorType, Kokkos::TeamPolicy<Properties...>,
     if ((arg_policy.scratch_size(0) +
          FunctorTeamShmemSize<FunctorType>::value(arg_functor,
                                                   arg_policy.team_size())) >
-        TeamPolicy<Kokkos::OpenMP>::scratch_size_max(0)) {
+        static_cast<size_t>(TeamPolicy<Kokkos::OpenMP>::scratch_size_max(0))) {
       std::stringstream error;
       error << "Requested too much scratch memory on level 0. Requested: "
             << arg_policy.scratch_size(0) +
@@ -411,7 +411,7 @@ class ParallelFor<FunctorType, Kokkos::TeamPolicy<Properties...>,
       Kokkos::abort(error.str().c_str());
     }
     if (arg_policy.scratch_size(1) >
-        TeamPolicy<Kokkos::OpenMP>::scratch_size_max(1)) {
+        static_cast<size_t>(TeamPolicy<Kokkos::OpenMP>::scratch_size_max(1))) {
       std::stringstream error;
       error << "Requested too much scratch memory on level 1. Requested: "
             << arg_policy.scratch_size(1)
