+++ core/src/OpenMP/Kokkos_OpenMP_Parallel_Reduce.hpp
@@ -532,7 +532,7 @@ class ParallelReduce<CombinedFunctorReducerType,
     if ((arg_policy.scratch_size(0) +
          FunctorTeamShmemSize<FunctorType>::value(
              m_functor_reducer.get_functor(), arg_policy.team_size())) >
-        TeamPolicy<Kokkos::OpenMP>::scratch_size_max(0)) {
+        static_cast<size_t>(TeamPolicy<Kokkos::OpenMP>::scratch_size_max(0))) {
       std::stringstream error;
       error << "Requested too much scratch memory on level 0. Requested: "
             << arg_policy.scratch_size(0) +
@@ -542,7 +542,7 @@ class ParallelReduce<CombinedFunctorReducerType,
       Kokkos::abort(error.str().c_str());
     }
     if (arg_policy.scratch_size(1) >
-        TeamPolicy<Kokkos::OpenMP>::scratch_size_max(1)) {
+        static_cast<size_t>(TeamPolicy<Kokkos::OpenMP>::scratch_size_max(1))) {
       std::stringstream error;
       error << "Requested too much scratch memory on level 1. Requested: "
             << arg_policy.scratch_size(1)
